Jumping working.

+++ com.unity.ml-agents.extensions/Runtime/Input/Adaptors/ButtonInputActionAdaptor.cs
@@ -14,14 +14,14 @@ namespace Unity.MLAgents.Extensions.Runtime.Input
         public void QueueInputEventForAction(InputAction action, InputControl control, ActionSpec actionSpec, in ActionBuffers actionBuffers)
         {
             var val = actionBuffers.DiscreteActions[0];
-            InputSystem.QueueDeltaStateEvent(control, (byte)val);
+            InputSystem.QueueDeltaStateEvent(control, (float)val);
         }
 
         public void WriteToHeuristic(InputAction action, in ActionBuffers actionBuffers)
         {
             var discreteActions = actionBuffers.DiscreteActions;
-            var val = action.ReadValue<byte>();
-            discreteActions[0] = val;
+            var val = action.ReadValue<float>();
+            discreteActions[0] = (int)val;
         }
     }
 }
